+++ R/model_refinement.R
@@ -38,11 +38,11 @@
 #'              data = MTPL)
 #' sev <- glm(amount ~ bm + zip, weights = nclaims,
 #'             family = Gamma(link = "log"),
-#'             data = MTPL %>% filter(amount > 0))
+#'             data = MTPL |> filter(amount > 0))
 #'
 #' # Add predictions for freq and sev to data, and calculate premium
-#' premium_df <- MTPL %>%
-#'    add_prediction(freq, sev) %>%
+#' premium_df <- MTPL |>
+#'    add_prediction(freq, sev) |>
 #'    mutate(premium = pred_nclaims_freq * pred_amount_sev)
 #'
 #' # Restrictions on risk factors for region (zip)
@@ -53,8 +53,8 @@
 #'             family = Gamma(link = "log"), data = premium_df)
 #'
 #' # Fit restricted model
-#' burn_rst <- burn %>%
-#'   restrict_coef(., zip_df) %>%
+#' burn_rst <- burn |>
+#'   restrict_coef(restrictions = zip_df) |>
 #'   update_glm()
 #'
 #' # Show rating factors
@@ -187,20 +187,20 @@ restrict_coef <- function(model, restrictions) {
 #' clusters_freq <- construct_tariff_classes(age_policyholder_frequency)
 #'
 #' # Add clusters to MTPL portfolio
-#' dat <- MTPL %>%
-#'   mutate(age_policyholder_freq_cat = clusters_freq$tariff_classes) %>%
-#'   mutate(across(where(is.character), as.factor)) %>%
+#' dat <- MTPL |>
+#'   mutate(age_policyholder_freq_cat = clusters_freq$tariff_classes) |>
+#'   mutate(across(where(is.character), as.factor)) |>
 #'   mutate(across(where(is.factor), ~biggest_reference(., exposure)))
 #'
 #' # Fit frequency and severity model
 #' freq <- glm(nclaims ~ bm + age_policyholder_freq_cat, offset = log(exposure),
 #'  family = poisson(), data = dat)
 #' sev <- glm(amount ~ bm + zip, weights = nclaims,
-#'  family = Gamma(link = "log"), data = dat %>% filter(amount > 0))
+#'  family = Gamma(link = "log"), data = dat |> filter(amount > 0))
 #'
 #' # Add predictions for freq and sev to data, and calculate premium
-#' premium_df <- dat %>%
-#'   add_prediction(freq, sev) %>%
+#' premium_df <- dat |>
+#'   add_prediction(freq, sev) |>
 #'   mutate(premium = pred_nclaims_freq * pred_amount_sev)
 #'
 #' # Fit unrestricted model
@@ -210,17 +210,17 @@ restrict_coef <- function(model, restrictions) {
 #'                          data = premium_df)
 #'
 #' # Impose smoothing and create figure
-#' burn_unrestricted %>%
+#' burn_unrestricted |>
 #'   smooth_coef(x_cut = "age_policyholder_freq_cat",
 #'               x_org = "age_policyholder",
-#'               breaks = seq(18, 95, 5)) %>%
+#'               breaks = seq(18, 95, 5)) |>
 #'   autoplot()
 #'
 #' # Impose smoothing and refit model
-#' burn_restricted <- burn_unrestricted %>%
+#' burn_restricted <- burn_unrestricted |>
 #'   smooth_coef(x_cut = "age_policyholder_freq_cat",
 #'               x_org = "age_policyholder",
-#'               breaks = seq(18, 95, 5)) %>%
+#'               breaks = seq(18, 95, 5)) |>
 #'   update_glm()
 #'
 #' # Show new rating factors
@@ -378,8 +378,8 @@ print.smooth <- function(x, ...) {
 #' freq <- glm(nclaims ~ bm + zip, weights = power, family = poisson(),
 #'  data = MTPL)
 #' zip_df <- data.frame(zip = c(0,1,2,3), zip_rst = c(0.8, 0.9, 1, 1.2))
-#' freq %>%
-#'   restrict_coef(., zip_df) %>%
+#' freq |>
+#'   restrict_coef(restrictions = zip_df) |>
 #'   autoplot()
 #'
 #' @export