+++ R/zzz.R
@@ -1,5 +1,5 @@
 globalVariables(c("n", "y", "x", "predicted", "lwr_95", "upr_95", "frequency",
-                  ".SD", ".N", ".", "..density..", "avg_claimsize",
+                  ".SD", ".N", ".", "density", "avg_claimsize",
                   "avg_premium", ":=", "difference", "maanden", "maanden_tot",
                   "id", "..count..", "level", "Coef", "type", "fm_no_offset",
                   "start_", "estimate", "end_", "col1", "yhat", "breaks_min",