+++ NEWS.md
@@ -1,5 +1,8 @@
 # insurancerating 0.7.4
 
+* `bootstrap_rmse()` now uses `after_stat(density)` instead of the deprecated dot-dot notation
+* `custom_theme` in `autoplot.univariate()` is added to customize the theme
+
 # insurancerating 0.7.3
 
 * `autoplot.univariate()` now generates a plot even when there are missing values in the rows